% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{explain.condition}
\alias{explain.condition}
\title{Explain the fbsmt_grade field}
\usage{
\method{explain}{condition}(field)
}
\arguments{
\item{field}{Field to explain}
}
\description{
Explain the fbsmt_grade field
}
