% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{explain.fbsmt_grade}
\alias{explain.fbsmt_grade}
\title{Explain the fbsmt_grade field}
\usage{
\method{explain}{fbsmt_grade}(field)
}
\arguments{
\item{field}{Field to explain}
}
\description{
Explain the fbsmt_grade field
}
