% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{explain.grade}
\alias{explain.grade}
\title{Explain the present_use field}
\usage{
\method{explain}{grade}(field)
}
\arguments{
\item{field}{Field to explain}
}
\description{
Explain the present_use field
}
