% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{explain.join_status}
\alias{explain.join_status}
\title{Explain the join_status field}
\usage{
\method{explain}{join_status}(field)
}
\arguments{
\item{field}{Field to explain}
}
\description{
Explain the join_status field
}
