% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kingco_submarkets}
\alias{kingco_submarkets}
\title{Custom residential submarkets in King County WA}
\format{
A \code{"data.frame"} with 95 rows and 2 variables
\describe{
  \item{area}{Assessment area or zone}
  \item{submarket}{Custom submarket data}
}
}
\source{
Author created
}
\usage{
data(kingco_submarkets)
}
\description{
Join to the 'area' field in `kingco_sales`
}
\keyword{datasets}
