% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{utilAddPinx}
\alias{utilAddPinx}
\title{Add Pinx field to data}
\usage{
utilAddPinx(x_df, condoComp = FALSE)
}
\arguments{
\item{x_df}{Dataset to add pinx to}

\item{condoComp}{[FALSE] Is this condoComp data?}
}
\description{
Add Pinx field to data
}
