% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{utilLeadZeros}
\alias{utilLeadZeros}
\title{Add leading zeros}
\usage{
utilLeadZeros(x_nbr, max_char = 6)
}
\arguments{
\item{x_nbr}{Number to add zeros to}

\item{max_char}{[6] Maximum number of characters}
}
\description{
Add leading zeros
}
