% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kingco_sales}
\alias{kingco_sales}
\title{King County Sales 1999 to 2019}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 480011 rows and 46 columns.}
\source{
King County Assessor: http://info.kingcounty.gov/assessor/DataDownload/
}
\usage{
data(kingco_sales)
}
\description{
King County residential transactions of single family and townhomes.
}
\keyword{datasets}
